+++ core-ng/src/main/java/core/framework/impl/web/site/WebSecurityInterceptor.java
@@ -14,8 +14,9 @@ import static core.framework.http.ContentType.TEXT_HTML;
 /**
  * @author neo
  */
-public final class WebSecurityInterceptor implements Interceptor {    // refer to https://www.owasp.org/index.php/OWASP_Secure_Headers_Project#tab=Headers
-    public String contentSecurityPolicy = "default-src 'self'; img-src 'self' data:; object-src 'none'; frame-src 'none';";
+public final class WebSecurityInterceptor implements Interceptor {  // refer to https://www.owasp.org/index.php/OWASP_Secure_Headers_Project#tab=Headers
+    // example: "default-src 'self'; img-src 'self' data:; object-src 'none'; frame-src 'none';"
+    public String contentSecurityPolicy;
 
     @Override
     public Response intercept(Invocation invocation) throws Exception {
@@ -33,7 +34,9 @@ public final class WebSecurityInterceptor implements Interceptor {    // refer t
         response.header("Strict-Transport-Security", "max-age=31536000");
         response.contentType().ifPresent(contentType -> {
             if (TEXT_HTML.mediaType.equals(contentType.mediaType)) {
-                response.header("Content-Security-Policy", contentSecurityPolicy);
+                if (contentSecurityPolicy != null) {
+                    response.header("Content-Security-Policy", contentSecurityPolicy);
+                }
                 response.header("X-XSS-Protection", "1; mode=block");       // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-XSS-Protection
             }
             response.header("X-Content-Type-Options", "nosniff");